refactor(JOE-942): extract task-dispatch, delegation, run-helpers
Split remaining high-volume work-store domains into acyclic leaves:
task-dispatch receipts/acquisitions, delegation progress helpers, and
run lifecycle/lease/abort helpers. Ratchet façade LOC budget 3950→3000
(~2954 lines).
