+++ products/gateway/docs/development/module-boundary-budget.json
@@ -740,9 +740,9 @@
  "fileLocBudgets": [
   {
    "path": "src/work-store.ts",
-   "maxLines": 3950,
+   "maxLines": 3000,
    "owner": "work_store",
-   "note": "Hard LOC budget (JOE-919/JOE-942); current ~3903 after task-helpers + human-gates extract; toward 2000"
+   "note": "Hard LOC budget (JOE-919/JOE-942); current ~2954 after task-dispatch + delegation + run-helpers extract; toward 2000"
   },
   {
    "path": "src/environments.ts",
